refactor(page): extract printing-monitor types + support snapshot helpers (phase 1, batch 3)
Move the printing-monitor type definitions into
features/printing/printingMonitorTypes.ts and the cached
support/kickstand collection-snapshot helpers into
features/supports/supportSnapshotHelpers.ts. Verbatim move, call sites
re-import by name; no behavior change.

page.tsx: 22,483 -> 22,356 lines. Verified via `tsc --noEmit`.
